// src/types/PatientDetailsTypes.ts
export interface PatientDetails {
  Appointment_Date: string;
  Appointment_Id: string;
  Appointment_Notes: string;
  Appointment_Status: string;
  Appointment_Time: string;
  Doctor_Name: string;
  Doctor_id: string;
  Patient_Id: string;
  Patient_Mobile: string;
  Patient_Name: string;
  Patient_Photo: string;
  Patient_country_code: string;
  Patient_email: string;
  Patient_mobile: string;
  Patient_phone: string;
  patient_code: string;
}


export interface PatientDataProps {
  Dob: string
  File_No: string,
  Name: string,
  Patient_Id: string,
  Photo: string,
  address: string,
  age: string,
  bloodgroup: string,
  city: string,
  country_code: string,
  email: string,
  gender: string,
  id: string,
  mobile: string,
  phone: string,
  pin: string
  patient_code: string
}